added profile statistics

+++ app/src/main/java/de/kniederelz/pawplan/tracking/repositories/session/domain/WalkingTrackerSessionExtensions.kt
@@ -2,10 +2,10 @@ package de.kniederelz.pawplan.tracking.repositories.session.domain
 
 import android.location.Location
 
-fun WalkingTrackerSession.getDistance(): Float {
-    if (locations.size < 2) return 0f
+fun WalkingTrackerSession.getDistance(): Double {
+    if (locations.size < 2) return 0.0
 
-    var total = 0f
+    var total = 0.0
     val result = FloatArray(1)
 
     for (i in 0 until locations.lastIndex) {
